Missed throwing feature.
Adds a more human like behaviour to the pokemon_catch_worker, by
reducing the 100% hit rate the bot had on every throw. This will add a
hit_rate parameter that will allow the bot to miss some shots just like
a human. Also randomises a bit the waiting times so it's not always the
same.